ENH: combined more grid utilities
Combined more of the grid utilities into a single file.
