+++ codebase/superdarn/src.bin/tk/tool/make_fov.1.0/grid_utils.c
@@ -1,5 +1,5 @@
-/* make_grid.c
-   ===========
+/* grid_utils.c
+   ============
 */
 
 /*
@@ -33,11 +33,12 @@
 #include <zlib.h>
 
 #include "rtypes.h"
+#include "rtime.h"
 #include "scandata.h"
 
 
 /**
- * Exclude beams that are not part of a scan
+ * Exclude beams that arxe not part of a scan
  **/
 int exclude_outofscan(struct RadarScan *ptr) {
 
@@ -106,5 +107,82 @@ double slant_range(int frang, int rsep,
 
     /* Return the calculated slant range [km] */
     return (lagfr-rxris+(range_gate-1)*smsep+range_edge)*0.15;
+
+}
+
+int set_stereo_channel(char *chnstr)
+{
+  int channel = 0;
+
+  if(chnstr != NULL)
+    {
+      if(tolower(chnstr[0]) == 'a') channel = 1;
+      else if(tolower(chnstr[0]) == 'b') channel = 2;
+      else
+	fprintf(stderr, "Unknown Stereo channel %s, leaving unset\n", chnstr);
+    }
+
+  return(channel);
+}
+
+int set_fix_channel(char *chnstr_fix)
+{
+  int channel_fix=-1;
+
+  if(chnstr_fix != NULL)
+    {
+      if(tolower(chnstr_fix[0]) == 'a') channel_fix = 1;
+      else if(tolower(chnstr_fix[0]) == 'b') channel_fix = 2;
+      else if(tolower(chnstr_fix[0]) == 'c') channel_fix = 3;
+      else if(tolower(chnstr_fix[0]) == 'd') channel_fix = 4;
+      else
+	fprintf(stderr, "Unknown fix channel %s, leaving unset\n", chnstr_fix);
+    }
+
+  return(channel_fix);
+}
+
+/**
+ * Converts an input date from YYYYMMDD format to an epoch time in number of
+ * seconds since 00:00 UT on January 1, 1970.
+ **/
+double strdate(char *text) {
+
+    double tme;
+    int val;
+    int yr,mo,dy;
+
+    /* Calculate day, month, and year from YYYYMMDD format date */
+    val=atoi(text);
+    dy=val % 100;
+    mo=(val / 100) % 100;
+    yr=(val / 10000);
+
+    /* If only 2-digit year provided then assume it was pre-2000 */
+    if (yr<1970) yr+=1900;
+
+    /* Calculate epoch time of input year, month, and day */
+    tme=TimeYMDHMSToEpoch(yr,mo,dy,0,0,0);
+
+    /* Return epoch time in number of seconds since 00:00UT on January 1, 1970 */
+    return tme;
+
+}
+
+
+/**
+ * Converts an input time from HHMM format to number of seconds.
+ **/
+double strtime(char *text) {
+
+    int hr,mn;
+    int i;
+
+    for (i=0;(text[i] !=':') && (text[i] !=0);i++);
+    if (text[i]==0) return atoi(text)*3600L;
+    text[i]=0;
+    hr=atoi(text);
+    mn=atoi(text+i+1);
+    return hr*3600L+mn*60L;
 
 }
